pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atria_core"
version = "0.0.0"                                                                                                    
description = "atria is a PyTorch-based toolkit designed for training/testing machine/deep learning models at scale."
readme = "README.md"
requires-python = "==3.11.*"
license = { file = "LICENSE.txt" }
authors = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]


[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
cpu = [
  "torch==2.1.2",
  "torchvision==0.16.2",
]
cu121 = [
  "torch==2.1.2",
  "torchvision==0.16.2",
]
test = ["coverage"]

[tool.coverage.report]
skip_covered = true
show_missing = true
precision = 2
exclude_lines = [
    'def __repr__',
    'pragma= no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING=',
    'if typing.TYPE_CHECKING=',
    '@overload',
    '@typing.overload',
    '\(Protocol\)=$',
    'typing.assert_never',
    'assert_never',
    'if __name__ == "__main__":',
]

## Linter and formatter
[tool.ruff]
# cover and extend the default config in https=//docs.astral.sh/ruff/configuration/
extend-exclude = ["alembic"]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
    "B904",   # Allow raising exceptions without from e, for HTTPException
    "COM819", # Trailing comma prohibited
    "D100",   # Missing docstring in public module(file)
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "E201",   # Whitespace after '('
    "E202",   # Whitespace before ')'
    "E203",   # Whitespace before '='
    "E221",   # Multiple spaces before operator
    "E241",   # Multiple spaces after ','
    "E251",   # Unexpected spaces around keyword / parameter equals
    "W291",   # Trailing whitespace
    "W293",   # Blank line contains whitespace
]

isort = { combine-as-imports = true,  split-on-trailing-comma = false }

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[project.urls]
"Homepage" = "https://github.com/saifullah3396/atria_core/"
"Bug Reports" = "https://github.com/saifullah3396/atria_core/"
"Source" = "https://github.com/saifullah3396/atria_core/"

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests"]
# Set additional command line options for pytest
# Ref: https://docs.pytest.org/en/stable/reference/reference.html#command-line-flags
addopts = "-rXs --strict-config --strict-markers --tb=short --import-mode=importlib"
xfail_strict = true         # Treat tests that are marked as xfail but pass as test failures
# filterwarnings = ["error"]  # Treat all warnings as errors
asyncio_default_fixture_loop_scope = "session"


[tool.uv]
managed = true
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu121" },
  ],
]
[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu"},
  { index = "pytorch-cu121", extra = "cu121" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu"},
  { index = "pytorch-cu121", extra = "cu121" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[dependency-groups]
all = [
    "codename==1.1",
    "coloredlogs==15.0.1",
    "gotrue==2.12.0",
    "lazy_loader==0.4",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "Pillow==10.0.0",
    "polyfactory>=2.21.0",
    "pytorch_ignite==0.5.0.post2",
    "rich==14.0.0",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "beautifulsoup4==4.12.2",
    "supabase>=2.7.4",
    "imagesize>=1.4.1",
    "ipykernel>=6.29.5",
    "pyarrow>=20.0.0",
    "factory-boy>=3.3.3",
    "ruff>=0.12.2",
    "ipykernel>=6.29.5",
    "factory-boy>=3.3.3",
    "bs4>=0.0.2",
    "types-beautifulsoup4",
    "lxml>=6.0.0",
]
dev = [
    "codename==1.1",
    "coloredlogs==15.0.1",
    "gotrue==2.12.0",
    "lazy_loader==0.4",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "Pillow==10.0.0",
    "polyfactory>=2.21.0",
    "pytorch_ignite==0.5.0.post2",
    "rich==14.0.0",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "beautifulsoup4==4.12.2",
    "supabase>=2.7.4",
    "imagesize>=1.4.1",
    "ipykernel>=6.29.5",
    "pyarrow>=20.0.0",
    "factory-boy>=3.3.3",
    "ruff>=0.12.2",
    "ipykernel>=6.29.5",
    "factory-boy>=3.3.3",
    "bs4>=0.0.2",
]
raw-types = [
    "codename==1.1",
    "coloredlogs==15.0.1",
    "gotrue==2.12.0",
    "lazy_loader==0.4",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "Pillow==10.0.0",
    "rich==14.0.0",
    "beautifulsoup4==4.12.2",
    "imagesize>=1.4.1",
    "pyarrow>=20.0.0",
    "bs4>=0.0.2",
]

[tool.mypy]
disable_error_code = ["import-untyped"]
